requirements.txt
psycopg2
click>=8.0.1
GitPython